+++ src/main/java/com/gildedgames/util/ui/util/TextBox.java
@@ -35,7 +35,7 @@ public class TextBox extends GuiFrame
 		int i = 0;
 		int textHeight = 0;
 
-		double halfWidth = this.getDim().width() / 2;
+		float halfWidth = this.getDim().width() / 2;
 		
 		for (Text t : this.text)
 		{
